+++ shopify_sdk/gql/queries.py
@@ -15,6 +15,9 @@ locations = VersionedGraphQLSymbol("queries", "locations")
 bulkOperation = VersionedGraphQLSymbol("queries", "bulkOperation")
 productSetOperation = VersionedGraphQLSymbol("queries", "productSetOperation")
 productByIdentifier = VersionedGraphQLSymbol("queries", "productByIdentifier")
+draftProductByIdentifier = VersionedGraphQLSymbol(
+    "queries", "draftProductByIdentifier"
+)
 metafieldDefinition = VersionedGraphQLSymbol("queries", "metafieldDefinition")
 
 __all__ = [
@@ -30,5 +33,6 @@ __all__ = [
     "bulkOperation",
     "productSetOperation",
     "productByIdentifier",
+    "draftProductByIdentifier",
     "metafieldDefinition",
 ]